Servicos recebendo float e convertendo int

+++ models/requests/service_update_data.py
@@ -6,4 +6,4 @@ from pydantic import BaseModel
 class ServiceUpdateData(BaseModel):
     name: Optional[str] = None
     description:  Optional[str] = None
-    base_price: Optional[int] = None
+    base_price: Optional[float] = None
